// src/models/LocationModel.ts
import mongoose, { Model, Schema } from "mongoose";
import { Location } from "@/types";

const LocationSchema: Schema = new Schema({
  // The CUID is a unique identifier that is generated by the cuid package
  // Is usally a good idea to use a public identifier for your documents instead of the _id field
  // Should be indexed for faster lookups
  cuid: { type: String, required: true, unique: true, index: true },
  name: { type: String, required: true },
  latitude: { type: Number, required: true },
  longitude: { type: Number, required: true },
  address: { type: String, required: true },
});

export const LocationModel: Model<Location> =
  mongoose.models.Location ||
  mongoose.model<Location>("Location", LocationSchema);